incomplete(lf_pager): lf_pager doesn't work when calling from windows
The lf_pager script does not work when lf is started on windows even
though it is calling the shell script version. It seems that there is an
issue with bat using a pager as it causes the folling error:

```
0 [main] less 3066 dtable::stdio_init: couldn't make stderr distinct from stdout, Win32 error 6
```

This does not happen when lf is started in git bash/zsh which is weird
as both shells should setup git bash as the running shell in windows.

This pathc only addresses the detection under powershell and cmd but
does not fix the pager or previews.